add ManageListsDialog to add and remove media to/from lists

+++ feature/account/data/src/main/java/com/buntupana/tmdb/feature/account/data/remote_data_source/raw/ListDetailRaw.kt
@@ -30,7 +30,7 @@ data class ListDetailRaw(
     @SerialName("public")
     val public: Boolean,
     @SerialName("revenue")
-    val revenue: Int?,
+    val revenue: Long?,
     @SerialName("runtime")
     val runtime: Int,
     @SerialName("sort_by")
